# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Odoo Cafe POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ── PAYMENT METHODS ──────────────────────────────────────────────────────────
...
  /tables:
    get:
      operationId: listTables
      tags: [tables]
      summary: List all tables
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/{id}:
    patch:
      operationId: updateTable
      tags: [tables]
      summary: Update table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ── COUPONS ──────────────────────────────────────────────────────────────────
...
  /reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Get report summary metrics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, all]
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: sessionId
          in: query
          schema:
            type: integer
        - name: productId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"
  /reports/sales-trend:
    get:
      operationId: getSalesTrend
      tags: [reports]
      summary: Sales trend over time
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, all]
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: sessionId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"
  /reports/top-products:
    get:
      operationId: getTopProducts
      tags: [reports]
      summary: Top selling products
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, all]
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /reports/top-categories:
    get:
      operationId: getTopCategories
      tags: [reports]
      summary: Top categories by revenue
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, all]
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopCategory"
  /reports/top-orders:
    get:
      operationId: getTopOrders
      tags: [reports]
      summary: Top orders by value
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, all]
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, categoryId, price, unit, tax, sendToKitchen]
      properties:
        id:
          type: integer
        name:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]
        price:
          type: number
        unit:
          type: string
        tax:
          type: number
        description:
          type: ["string", "null"]
        sendToKitchen:
          type: boolean
    ProductInput:
      type: object
      required: [name, categoryId, price, unit, tax]
      properties:
        name:
          type: string
        categoryId:
          type: integer
        price:
          type: number
        unit:
          type: string
        tax:
          type: number
        description:
          type: string
        sendToKitchen:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        categoryId:
          type: integer
        price:
          type: number
        unit:
          type: string
        tax:
          type: number
        description:
          type: string
        sendToKitchen:
          type: boolean

    # ── PaymentMethod ─────────────────────────────────────────────────────────
...
    Table:
      type: object
      required: [id, floorId, number, seats, active]
      properties:
        id:
          type: integer
        floorId:
          type: integer
        floorName:
          type: ["string", "null"]
        number:
          type: integer
        seats:
          type: integer
        active:
          type: boolean
        hasActiveOrder:
          type: boolean
    TableInput:
      type: object
      required: [floorId, number, seats]
      properties:
        floorId:
          type: integer
        number:
          type: integer
        seats:
          type: integer
        active:
          type: boolean
    TableUpdate:
      type: object
      properties:
        floorId:
          type: integer
        number:
          type: integer
        seats:
          type: integer
        active:
          type: boolean

    # ── Coupon ────────────────────────────────────────────────────────────────
...
    CartLine:
      type: object
      required: [productId, qty, unitPrice]
      properties:
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        qty:
          type: number
        unitPrice:
          type: number
        lineTotal:
          type: number
        productDiscount:
          type: ["number", "null"]
        productDiscountLabel:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, number, lines, subtotal, tax, discountTotal, total, status, createdAt]
      properties:
        id:
          type: integer
        number:
          type: string
        tableId:
          type: ["integer", "null"]
        tableNumber:
          type: ["integer", "null"]
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        lines:
          type: array
          items:
            $ref: "#/components/schemas/CartLine"
        subtotal:
          type: number
        tax:
          type: number
        discountTotal:
          type: number
        discountLabel:
          type: ["string", "null"]
        total:
          type: number
        status:
          type: string
          enum: [Draft, Paid, Cancelled]
        paymentMethodId:
          type: ["integer", "null"]
        paymentRef:
          type: ["string", "null"]
        amountPaid:
          type: ["number", "null"]
        sentToKitchen:
          type: boolean
        employeeId:
          type: ["integer", "null"]
        sessionId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    ReportSummary:
      type: object
      required: [totalOrders, revenue, avgOrderValue]
      properties:
        totalOrders:
          type: integer
        revenue:
          type: number
        avgOrderValue:
          type: number
    TrendPoint:
      type: object
      required: [label, revenue, orders]
      properties:
        label:
          type: string
        revenue:
          type: number
        orders:
          type: integer
    TopProduct:
      type: object
      required: [productId, productName, qty, revenue]
      properties:
        productId:
          type: integer
        productName:
          type: string
        qty:
          type: number
        revenue:
          type: number
    TopCategory:
      type: object
      required: [categoryId, categoryName, categoryColor, revenue]
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryColor:
          type: string
        revenue:
          type: number